requirements.txt
easygui==0.98.3
numpy==1.26.4
opencv_contrib_python==4.9.0.80
pycups==2.0.1
